🌟 add typing feature

+++ telegram-bot/src/main/kotlin/com/theapache64/cs/models/TelegramCallbackQuery.kt
@@ -42,7 +42,7 @@ data class TelegramCallbackQuery(
             @SerializedName("from")
             val from: From,
             @SerializedName("message_id")
-            val messageId: Int, // 61
+            val messageId: Long, // 61
             @SerializedName("reply_markup")
             val replyMarkup: ReplyMarkup,
             @SerializedName("text")
